EP-3042: smallint integeriksi

+++ eperusteet/eperusteet-service/src/main/resources/db/migration/V0_20220310154400__koto_tyoelama_uudet_kentat.sql
@@ -1,10 +1,10 @@
 alter table koto_taitotaso add column opiskelijanTyoelamataidot_id bigint;
-alter table koto_taitotaso add column tyoelamaOpintoMinimiLaajuus smallint;
-alter table koto_taitotaso add column tyoelamaOpintoMaksimiLaajuus smallint;
+alter table koto_taitotaso add column tyoelamaOpintoMinimiLaajuus integer;
+alter table koto_taitotaso add column tyoelamaOpintoMaksimiLaajuus integer;
 
 alter table koto_taitotaso_aud add column opiskelijanTyoelamataidot_id bigint;
-alter table koto_taitotaso_aud add column tyoelamaOpintoMinimiLaajuus smallint;
-alter table koto_taitotaso_aud add column tyoelamaOpintoMaksimiLaajuus smallint;
+alter table koto_taitotaso_aud add column tyoelamaOpintoMinimiLaajuus integer;
+alter table koto_taitotaso_aud add column tyoelamaOpintoMaksimiLaajuus integer;
 
 alter table koto_taitotaso
     add constraint FK_4h76eyjxiji0c4kejakptk1iu
